implement customizeCloudInit for classic images
If a cloud-init confguration is given (meta-data, user-data or
netowrk-config, it is writen in the chroot in
/var/lib/cloud/seed/nocloud. This is a supported (yet not documented)
way of configuring the NoCloud datasource.
The datasource list is set in /etc/cloud/cloud.cfg.d/90_dpkg.cfg to
NoCloud only.
This function get called only if a cloud-init config is present.