Fix directory reuse in gadget build test
TestBuildGadgetTreeDirectory.test_build_gadget_tree_directory aims
to verify that building a gadget works using the "directory" method and
that the result has a gadget.yaml in the right place. It was building
first from local data, then from the github hosted pc-gadget repository.

Unfortunately, this resulted in the second gadget being checked out
over the first one, and since the first one used "makefile" and the
second one "Makefile", both builds were really using "makefile". This
used to pass because gadget.yaml was present at the top of the gadget
repository, but now that it's generated as part of the build, Makefile
must really be used, and it wasn't sane to reuse the same directory for
two gadget builds anyway.

Since the first build wasn't providing any additional test coverage,
just run the second build.