Run the spread tests using google backend in self-hosted runners
This change allowes to run the spread tests in slef-hosted runners in
the canonical env and also using the google backend instead of lxd.

The spread job includes the initial cleanup, the tests execution and the
last step used to make sure all the vms are deleted in google cloud.