Add --import-proxy-asserts option for `snap` subcommand
--import-proxy-asserts accepts a Store Proxy URL as a parameter.

Using that URL, ubuntu-image then fetches the store assertion,
account-key assertion, and account assertion, needed to validate the
store on a device, and appends it to `model-etc`, so that each of the
assertions is imported into snapd's assertion store on first-boot.

This allows device images to be built such that they can point to a
Store Proxy out of the box.

Signed-off-by: Joseph Brock <joseph.brock@canonical.com>