i/{statemachine,helper}: Use the sparse option on tar
Uncompressed archive benefit the most out of this improvement.
Compressed archives are not smaller (only tested on small archives, this might not be fully representative) but according to the doc it should be at least faster to generate.
Also rework test to check the size so we could detect if the `--sparse` flag is removed in a future rework.

Signed-off-by: Paul Mars <paul.mars@canonical.com>