i/ppa: Create gpg tmp dir outside of workdir
Rework createTmpGPGDir so the temp dir provided to GPG is in the OS-chosen temp directory. This should very likely always be a short directory and so should not be an issue for the socket name used to communicate with dirmngr.

Also make sure we "hook" the call to os.MkdirTemp in tests so the temp dir is in fact created in the temp directory dedicated to the test (and cleaned up at the end). This *may* break if tests are run from of deep directory.

Fixes: LP: #2057885
Signed-off-by: Paul Mars <paul.mars@canonical.com>