Fix build components with legacy sources.list
During build, the Rootfs definitions should be used for APT sources, while the
Customization definitions should be used in the final image. For the legacy APT
sources generation, we only followed this logic for the pocket value, not for
components.

This might also make sense for mirror, but that's not a standalone
Customization option.